Renamed Unit to EnergyUnit.
